# Simpler iterator interface