--- a/common/kv.go
+++ b/common/kv.go
@@ -4,9 +4,3 @@ type KV struct {
 	Key   []byte
 	Value []byte
 }
-
-type KVV struct {
-	Key     []byte
-	Value   []byte
-	Version uint64
-}