--- a/iteration/chaining_iter.go
+++ b/iteration/chaining_iter.go
@@ -13,30 +13,15 @@ func NewChainingIterator(its []Iterator) *ChainingIterator {
 	return &ChainingIterator{iterators: its}
 }
 
-func (c *ChainingIterator) Current() common.KV {
-	return c.iterators[c.pos].Current()
-}
-
-func (c *ChainingIterator) Next() error {
-	if err := c.iterators[c.pos].Next(); err != nil {
-		return err
+func (c *ChainingIterator) Next() (bool, common.KV, error) {
+	for ; c.pos < len(c.iterators); c.pos++ {
+		valid, kv, err := c.iterators[c.pos].Next()
+		if err == nil && !valid {
+			continue
+		}
+		return valid, kv, err
 	}
-	valid, err := c.iterators[c.pos].IsValid()
-	if err != nil {
-		return err
-	}
-	if valid {
-		return nil
-	}
-	c.pos++
-	return nil
-}
-
-func (c *ChainingIterator) IsValid() (bool, error) {
-	if c.pos >= len(c.iterators) {
-		return false, nil
-	}
-	return c.iterators[c.pos].IsValid()
+	return false, common.KV{}, nil
 }
 
 func (c *ChainingIterator) Close() {