--- a/iteration/gen_iter.go
+++ b/iteration/gen_iter.go
@@ -51,12 +51,22 @@ func (s *StaticIterator) AddKV(k []byte, v []byte) {
 }
 
 func (s *StaticIterator) Next() (bool, common.KV, error) {
-	if (s.hasValidOverride && !s.validOverRideValue) || s.pos >= len(s.kvs) {
+	if s.pos >= len(s.kvs) {
+		s.pos = -1
+	}
+	if (s.hasValidOverride && !s.validOverRideValue) || s.pos == -1 {
 		return false, common.KV{}, nil
 	}
 	v := s.kvs[s.pos]
 	s.pos++
 	return true, v, nil
+}
+
+func (s *StaticIterator) Current() common.KV {
+	if s.pos <= 0 {
+		return common.KV{}
+	}
+	return s.kvs[s.pos-1]
 }
 
 func (s *StaticIterator) Close() {