--- a/iteration/iter.go
+++ b/iteration/iter.go
@@ -1,10 +1,10 @@
 package iteration
 
-import "github.com/spirit-labs/tektite/common"
+import (
+	"github.com/spirit-labs/tektite/common"
+)
 
 type Iterator interface {
-	Current() common.KV
-	Next() error
-	IsValid() (bool, error)
+	Next() (bool, common.KV, error)
 	Close()
 }