solves https://github.com/spirit-labs/tektite/issues/82