@purplefox 2024-07-04
Hi @watson28 thanks for this!

There are a few conflicts after I merged the work on refactoring store. The store stuff has mainly moved to proc/storage.go