@watson28 2024-07-05
@purplefox I will wait for the merge of https://github.com/spirit-labs/tektite/pull/142 since it has additional changes to iterators.