@purplefox 2024-07-05
It's merged now. Hopefully won't be changes to iterators for a while :)