@purplefox 2024-07-05
Since MergingIterator is performance sensitive code, it would nice to see a benchmark which iterates through a bunch of keys using MergingIterator and compares performance before and after the changes.