@watson28 2024-07-08
@purplefox I addressed your comment:
-  I changed the Next() method to return an error
- I moved the logic in `dropNextKeysEqualsToLatest` to `readIterHeadOrNext`

Regarding:
>Since MergingIterator is performance-sensitive code, it would be nice to see a bench iterates through a bunch of keys using MergingIterator and compares performance before and after the changes.

It seems we have `merging_iter_bench_test.go` for this. I ran the benchmark, and thats is what I got:

This branch:
```
goos: darwin
goarch: arm64
pkg: github.com/spirit-labs/tektite/iteration
BenchmarkMergingIterator-10    	    5548	    183379 ns/op	   64355 B/op	    1003 allocs/op
PASS
ok  	github.com/spirit-labs/tektite/iteration	1.262s
```

main branch:
```
goos: darwin
goarch: arm64
pkg: github.com/spirit-labs/tektite/iteration
BenchmarkMergingIterator-10    	    3578	    305500 ns/op	     296 B/op	       3 allocs/op
PASS
ok  	github.com/spirit-labs/tektite/iteration	1.278s
```

It seems we now have an improvement in the "Nanoseconds per operation" (third column) but now we are allocating more memory (last two columns).

I will work on refactoring the code to consume less memory.