@purplefox 2024-07-08
I wonder if maintaining `iterHeads` is really necessary?