@purplefox 2024-07-08
> I will work on refactoring the code to consume less memory.

Ack, it should be possible to avoid creating new structs in the loops (kvv etc)