@watson28 2024-07-10
@purplefox 

this is the new benchmark with the latest changes

```
goos: darwin
goarch: arm64
pkg: github.com/spirit-labs/tektite/iteration
BenchmarkMergingIterator-10    	    6270	    189081 ns/op	     277 B/op	       3 allocs/op
PASS
```
Now we have better numbers in "Nanoseconds per operation" and "Number of bytes allocated per operation".

basically, the increase in memory was caused by the new functions  I created to improve the readability of the code (`readIterHead` and the 3 functions to log messages). New function calls means more memory stack allocated.  I inlined the code in the function as it originally was, and the memory usage was reduced.

The other change needed was to remove the `iterHeads` array. However, to do this, I have to bring back the `Current()` method to the iterator interface. This is needed in `merging_iter` since we have to keep a reference to the latest iter.Next() that is returned, in case that is not used in the current Next() call of `merging_iter`. Taking this example for the unit test:

```
it1: [2, 5, 9]
it2: [0, 7, 10]
it3: [4, 6, 8]
```

calling `mergin_iter.Next()` will call next() in the 3 iterators and choose the smallest key, it2 -> 0. But then we have advanced it1 and it3. so we need to have a way to keep the latest value of each item. Otherwise, we will lose the values `2` and `4` in the `Next()` call of `merging_iter`. That was achieved with `iterHeads` and now with the `Current()` method in the iterators.