@purplefox 2024-07-10
Thanks! I'll give this a proper look tomorrow morning.