@purplefox 2024-07-10
Just a thought: If we're going to re-introduce `Current()`, then it seems redundant for `Next()` to return a KV. I would suggest it should just return `(bool, error)`