@watson28 2024-07-11
>Just a thought: If we're going to re-introduce Current(), then it seems redundant for Next() to return a KV. I would suggest it should just return (bool, error)

@purplefox After looking at the code base, I noticed that most of the uses of iterators, except for `merging_iter,` satisfy their needs by only calling `Next()` and getting the value from the return value. It feels weird to change these and replace a single call with two function calls only because `merging_iter` needs it. I think the proper iterator interface should have the `Next()` method return the value and still provide `Current()` for those few cases where caching the last Next() Value is not an option for the client(`mergin_iter` case).

Let me know what you think.