@purplefox 2024-07-11
> Let me know what you think.

That's reasonable. I will merge.

Thanks for your efforts! Great piece of work