@purplefox 2024-07-11
I've had to revert this PR for now, as it seems to cause the continuous CI run to fail, e.g. https://github.com/spirit-labs/tektite/actions/runs/9896394329