@purplefox 2024-07-11
I would suggest running the failover integration tests in a loop locally. (There is a script `run_tests_in_loop.sh` in the root of the project which can help with that.
From past experience I would guess the problem is something related to the minCompactableVersion logic in the merging iterator, as this logic is exercised during failover.