@purplefox 2024-07-12
One way to see if there any behavioural differences between the old merging iterator and the new one (and this will probably be easier than debugging the failover tests on CI), would be to use a form of fuzz testing.

In a loop, generate a bunch of random keys and similar keys, with same and different versions and then run them through both the old and new merging iterators and verify that the output is the same.