@purplefox 2024-07-12
@watson28 Just to keep you informed - I have found a couple of bugs and will submit a patch in the next day or so.