@watson28 2024-07-16
>@watson28 Just to keep you informed - I have found a couple of bugs and will submit a patch in the next day or so.

@purplefox, thanks for reapplying the changes and the patch with the fix.